Added option to remove job ampersand from script
Made it so the command that this script runs need not have an & sign at the end of it. This cases all tasks on a on job list to run sequentially rather than in parallel. This is useful when running simultaneous tasks cause the node to run out of memory.

Currently only works for LISA systems!