dos step has been split into two commands
Have split up the dos script into dos (convert from 3 probability to 2 probability format) and dos2 (merge all dos files together). This was because the system tended to run out of memory when running the dos2 step for large files. I made it so the dos2 tasks are run sequentially on any node by removing the & sign after the command, preventing the system from running out of memory.