monomorphic markers now determined with all subjects
Monomorphic markers are determined by looking at which markers give NA p-values in the association analysis. Because I set to only retain a subset of subjects for the association analysis, more markers were monomorphic than should be. I modified the analysis step so that it runs 2 analyses: the 1st in just the --keep subjects (for the GWAS manhattan plots) and the 2nd in all subjects (for determining monomorphic markers).