Update Illumina loci-name rs-id annotations for imputation step
This script updates loci name to rs-ids. In the event of duplicates, the duplicate with the higher missingness rate is set to have '_duplicate' as a suffix. Multiple duplicates have '_duplicate_duplicate'. Currently works for up to 4 duplications of a single rs-id. Copying the duplication line will increase this number, but it shouldn't necessary.