Do not produce bg or bgs outputs
bg and bgs outputs are filtered on SNP missingness and MAF, which are ancestry dependent. Because mixed ancestries are in these files, these data are incorrect. These outputs are removed. This filtering can be done later at an analysis step anyway.

Only bgn files are retained (best guesss, no filtering)