--pheno used in postimp_navi_18 to curcumvent error
The dabg step currently takes the -9 phenotype of the .fam file and converts it to value 1 (for control). This was done to prevent an error in another program that didn't expect missing data in the .fam file. As a consequence, these subjects are stuck labeled as controls. I circumvent this error in the analysis by taking the QCed .fam file phenotype and supplying it to the postim command.