Platform guessing
The plague script generates a % correspondence between user data and sample array data. 

This script:
For each platform, there are two % metrics calculated. For each platform, the sum of the % metrics is taken. The platform with the highest sum is the assigned platform.

This is meant to supplant the perl code in the preimp_dirsub which does the same thing.